Cambios en las tablas

+++ database/migrations/2021_01_15_165840_create_paquetes_table.php
@@ -21,7 +21,7 @@ class CreatePaquetesTable extends Migration
             $table->string('description');
             $table->string('type');
             $table->double('price');
-            $table->double('stock');
+            $table->bigInteger('stock');
             $table->string('date');
             $table->timestamps();
         });
