+++ database/migrations/2021_01_15_205543_create_cruceros_table.php
@@ -21,7 +21,7 @@ class CreateCrucerosTable extends Migration
             $table->string('description');
             $table->string('type');
             $table->double('price');
-            $table->double('stock');
+            $table->bigInteger('stock');
             $table->string('date');
             $table->timestamps();
         });
